+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.1.8-preview.3] - 2019-01-17
+
+### Added
+* Suspend/Resume/Quit events support
+* Updated `com.unity.platforms` package to version `0.1.8-preview.1`.
+
+## [0.1.8-preview.1] - 2019-12-16
+
+### Changed
+* Re-implementing `Using static libs in XCode project instead of dynamic`
+* Fixed problem with resources which are not packed to Data files
+
+## [0.1.8-preview] - 2019-12-11
+
+### Changed
+* Bump com.unity.platforms version to 0.1.8-preview
+
 ## [0.1.7-preview.10] - 2019-12-11
 
 ### Changed